layer-shell: Improve configuration
We used to show the surface right away in order to trigger a commit,
but that creates the buffer before configuration and that breaks the
protocol.

Instead we manually send a commit after all the layer surface
attributes are set.  This will emit the changed signal that must be
interpreted by the compositor as its queue to configure the surface.