Convert image format to match the buffer format
The protocol demands that we set up a buffer and then capture the
output, this means the buffer format is fixed and sometimes didn't
match the QImage format.

As a matter of fact, when this code was written the QImage format was
Format_RGB32 but now it appears to be Format_RGBA8888_Premultiplied.

We better convert the final image format to the buffer format.