Improve the color picker protocol
A client-side generated serial will conflict as soon as a second client
is connected and generates the same number.

The protocol is now redesigned so that both pick requests will create
the object that in turn will emit an event with the picked color value.

The pick_interactively request no longer work on a specific output
because it didn't make any sense.