Layer shell integration
We currently use a QML library for layer shell on the client side that
doesn't integrate very well with the platform plugin.
Windows are made visible as soon as they are configured and when they
are hidden and shown again, they are not recreated.
Also, it's hard to handle popup and modal windows but with a shell
integration plugin we could hide all those details very well.

The idea is to make a LayerSurface interface exposed to clients
written with QML, then the shell integration plugin finds the
interface associated with the window and gets all the layer surface
attributes from it.
While the shell integration plugin depends on a private API and will
need to be rebuilt every time Qt is updated, the interface can be public.

This repository contains the LayerSurface interface.

Issue: #29